+++ src/main/java/nextstep/ladder/Main.java
@@ -1,5 +1,8 @@
 package nextstep.ladder;
 
+import nextstep.ladder.controller.LadderGame;
+import nextstep.ladder.view.LadderGameView;
+
 public class Main {
 
     public static void main(String[] args) {
