+++ src/main/java/nextstep/ladder/domain/Points.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import nextstep.UniformRandomBooleanProvider;
