+++ src/test/java/nextstep/ladder/ParticipantTest.java
@@ -1,5 +1,7 @@
 package nextstep.ladder;
 
+import nextstep.ladder.domain.Participant;
+import nextstep.ladder.domain.Participants;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
