+++ src/test/java/nextstep/ladder/PointTest.java
@@ -1,5 +1,7 @@
 package nextstep.ladder;
 
+import nextstep.ladder.domain.Point;
+import nextstep.ladder.domain.PointType;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
